+++ integration-tests/src/test/java/eu/xenit/alfred/telemetry/integrationtesting/AlfrescoPrometheusEndpointTest.java
@@ -37,9 +37,8 @@ public class AlfrescoPrometheusEndpointTest extends RestAssuredTest {
 
         String responseBody = response.body().asString();
         assertThat(responseBody, containsString("alfred_telemetry_registries_total"));
-        // Does not work with integrated Alfresco Prometheus registry
-//        assertThat(responseBody, containsString("application=\"alfresco\""));
-//        assertThat(responseBody, containsString(",host=\""));
+        assertThat(responseBody, containsString("application=\"alfresco\""));
+        assertThat(responseBody, containsString(",host=\""));
     }
 
 }